use the zip64 extended info only if the value is 0xFFFFFFFF
fixes #14

Relevant part of the spec is quoted below. There seem to be cases
where the size in local header is correct, but the one is zip64
extended header is zero. I don't see why the zip64 header is added in
the first place, since it's optional, anyhow, we will follow the spec
exactly and not use zip64 header unless the value is 0xFFFFFFFF.

If an archive is in ZIP64 format and the value in this field is
0xFFFFFFFF, the size will be in the corresponding 8 byte ZIP64
extended information extra field.  When encrypting the central
directory, if the local header is not in ZIP64 format and general
purpose bit flag 13 is set indicating masking, the value stored for
the uncompressed size in the Local Header will be zero.